+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme" >
         <activity
-            android:name=".OnBoardingActivity"
+            android:name=".MainPageActivity"
             android:label="@string/app_name" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -22,10 +22,10 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".UserPageActivity"
+            android:name=".GamePageActivity"
             />
         <activity
-            android:name=".AppointmentActivity"
+            android:name=".OnBoardingActivity"
             />
     </application>
