--- app/src/main/java/com/example/cardgame/cardgame/AppointmentActivity.java
@@ -1,14 +1,0 @@
-package com.example.cardgame.cardgame;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-
-public class AppointmentActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_userPage);
-    }
-    
-}
