+++ app/src/main/res/layout/activity_game_page.xml
