+++ app/src/main/res/layout/activity_onboarding.xml
